Collecting suitable breweries for travel.

+++ src/main/java/com/satalia/beer/model/Coordinates.java
@@ -18,10 +18,10 @@ public class Coordinates {
 	private Long breweryId;
 
 	@Column(name="Latitude")
-	private BigDecimal latitude;
+	private Double latitude;
 
 	@Column(name="Longitude")
-	private BigDecimal longitude;
+	private Double longitude;
 
 	public Long getId() {
 		return id;
@@ -39,19 +39,19 @@ public class Coordinates {
 		this.breweryId = breweryId;
 	}
 
-	public BigDecimal getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(BigDecimal latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public BigDecimal getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(BigDecimal longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
 }
